Move: patterns.js
