+++ src/modules/patterns.js
