+++ src/routes/rooms.js
@@ -4,7 +4,7 @@ const router = express.Router();
 
 const roomHandlers = require("../services/rooms");
 const validator = require("../middlewares/validator");
-const patterns = require("../db/patterns");
+const patterns = require("../modules/patterns");
 
 // 라우터 접근 시 로그인 필요
 router.use(require("../middlewares/auth"));
